% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{formatP}
\alias{formatP}
\title{A handy function for formatting p-values to nice percents for publication}
\usage{
formatP(p, min.sig = 0.001, max.sig = 99.99, sig.digs = 3)
}
\arguments{
\item{p}{A single number with p-values to be formatted.}

\item{min.sig, max.sig}{Numbers in (0, 100) defining when function should return a thing like '<0.01\%' or '>99.99\%' instead of number.
By default \strong{0.001} and \strong{99.99}.}

\item{sig.digs}{A number of significant digits to show. By default \strong{3}.}
}
\value{
A character vector with p-values formatted in a nice way
}
\description{
...
}
\details{
TODO: Allow vectorized input
TODO: Add a way to automatically format them in APA/NEJM formats
TODO: Add a way to skip a trailing 0 and add p/P in the beginning
TODO: Add option to sanitize input
}
\examples{
formatP(0.0026485635335) # -> "0.265\%"
formatP(5e-33) # -> "<0.001\%"
}
\seealso{
Other publishing-formatting: 
\code{\link{formatOR}()},
\code{\link{signif_sym}()}
}
\concept{publishing-formatting}
