% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_of_life.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Infix not in operator}
\usage{
lhs \%notin\% rhs
}
\arguments{
\item{lhs}{An object or vector}

\item{rhs}{A vector (a set)}
}
\value{
The vector of logical values being the negated \%in\% result
}
\description{
Function being basically the opposite of \%in\% operator. Strictly a QoL thing
}
\examples{
\dontrun{
 1 \%notin\% c(2, 1, 3) # -> F
 c(1, 5, 10) \%notin\% c(2,1,3) # -> F T T
}
}
\concept{QoL}
