% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{raporthtml}
\alias{raporthtml}
\title{That's a custom R Markdown html format for a custom template "raport-html".}
\usage{
raporthtml(
  custom_GA = FALSE,
  use_GA = FALSE,
  fig_width = 8,
  fig_height = 5,
  fig_caption = TRUE,
  lightbox = FALSE,
  thumbnails = FALSE,
  gallery = FALSE,
  toc_depth = 5,
  embed_fonts = TRUE,
  use_bookdown = FALSE,
  pandoc_args = NULL,
  md_extensions = NULL,
  mathjax = "rmdformats",
  highlight = NULL,
  default_style = c("light", "dark"),
  downcute_theme = c("default", "chaos"),
  ...
)
}
\arguments{
\item{custom_GA}{if TRUE, template should look for a custom \code{GA_header.html} file, instead if using my own with my own GA tag}

\item{use_GA}{if TRUE includes a Google Analytics header, FALSE by default}

\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{lightbox}{if TRUE, add lightbox effect to content images}

\item{thumbnails}{if TRUE display content images as thumbnails}

\item{gallery}{if TRUE and lightbox is TRUE, add a gallery navigation between images in lightbox display}

\item{toc_depth}{adjust table of contents depth}

\item{embed_fonts}{if TRUE, use local files for fonts used in the template. This leads to bigger files but ensures that these fonts are available. If FALSE they are downloaded from Google Web Fonts.}

\item{use_bookdown}{if TRUE, uses \code{\link[bookdown]{html_document2}} instead of \code{\link[rmarkdown]{html_document}}, thus providing numbered sections and cross references}

\item{pandoc_args}{arguments passed to the pandoc_args argument of rmarkdown \code{\link[rmarkdown]{html_document}}}

\item{md_extensions}{arguments passed to the md_extensions argument of rmarkdown \code{\link[rmarkdown]{html_document}}}

\item{mathjax}{set to NULL to disable Mathjax insertion}

\item{highlight}{syntax highlighting, forced to NULL as highlighting is done via prism.js}

\item{default_style}{specify default display style, "light" or "dark"}

\item{downcute_theme}{document template theme}

\item{...}{Additional function arguments passed to R Markdown \code{\link[rmarkdown]{html_document}}}
}
\value{
An custom R Markdown html output format.
}
\description{
That's a custom R Markdown html format for a custom template "raport-html".
}
\details{
It uses by default the \code{downcute} template from \code{rmdformats} library, has my
custom Google Analytics tag and has some preloaded libraries.

The parameters are copied from the \code{downcute} options from \code{juba}'s Github.
}
