% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{remove_diacritics}
\alias{remove_diacritics}
\title{Replaces Polish diacritic letters with their plain Latin equivalents}
\usage{
remove_diacritics(string)
}
\arguments{
\item{string}{A character vector to be transformed}
}
\value{
A character vector without Polish diacritic letters
}
\description{
Comes practical when some functions doesn't work well with UTF-8 characters. Like for example roxygen documentation generator ;-)
}
\examples{
remove_diacritics("Wis\u0142a i \u017C\u00F3\u0142\u0107") # -> "Wisla i zolc"
# Unfortunately Roxygen documentation can have non ASCII characters only as escape characters

}
