% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{signif_sym}
\alias{signif_sym}
\title{Changes p-value number to a significance symbol for decorative purposes}
\usage{
signif_sym(
  x,
  max_asterisks = 5,
  bonferoni = 1,
  NA_instead_of_error = F,
  include_dot = T,
  ns_sym = " "
)
}
\arguments{
\item{x}{A numeric vector of input probabilities}

\item{max_asterisks}{Number of asterisks (*) for the most significant results, default is 5. Cannot be smaller than 2. Set to 3 for APA/NEJM standards.}

\item{bonferoni}{Number of tests in a family of tests when applying the Bonferroni correction for multiple comparisons problem. Is 1 by default which is same as not applying any correction.
Use it only if the passed p-values are not corrected already.}

\item{NA_instead_of_error}{Logical, by default (FALSE) if the functions sees any probability outside <0,1> it raises an error. However this behaviour can be changed to return NAs instead.}

\item{include_dot}{Logical, default TRUE, should the dot symbol be used for borderline significance that is between 5\% and 10\%?}

\item{ns_sym}{Character (default ' '), the symbol used for the non-significant results. Set to 'ns' APA/NEJM standards.}
}
\value{
A character vector with significance symbols.
}
\description{
Changes p-value number to a significance symbol for decorative purposes
}
\examples{
x <- 3e-4
signif_sym(x)
}
