% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{signif_sym}
\alias{signif_sym}
\title{Changes p-value number to a significance symbol for decorative purposes}
\usage{
signif_sym(x)
}
\arguments{
\item{x}{A number vector with one element}
}
\value{
A character vector.
}
\description{
Changes p-value number to a significance symbol for decorative purposes
}
\examples{
x <- 3e-4
signif_sym(x)
}
